Make na new class cpkreader
